feat(fsd): 实现飞行计划到达检测与删除功能
- 新增到达机场数据和位置信息字段
- 实现飞行计划到达检测逻辑
- 自动删除已到达的飞行计划
- 优化飞行计划更新流程
- 修复模拟器服务器判断逻辑
- 增加机场数据范围类型为浮点数支持

+++ internal/interfaces/config/airport_data.go
@@ -11,7 +11,7 @@ type AirportData struct {
 	Lat          float64 `json:"lat"`
 	Lon          float64 `json:"lon"`
 	Alt          float64 `json:"alt"`
-	AirportRange int     `json:"airport_range"`
+	AirportRange float64 `json:"airport_range"`
 }
 
 func (config *AirportData) checkValid(_ log.LoggerInterface) *ValidResult {
